requirements.txt
# ============================================
# AGRI-SENTINEL — Production Dependencies
# Pinned versions for reproducible builds
# ============================================

fastapi==0.129.0
uvicorn[standard]==0.40.0
gunicorn
supabase==2.28.0
python-multipart
python-dotenv
google-generativeai==0.8.6
jinja2==3.1.6
pillow==12.1.1
httpx==0.28.1
itsdangerous==2.2.0
pydantic[email]==2.12.5
requests==2.32.5
